Update Android build.
The Android build was broken when GLAD was added as a loader.
This CL includes a number of fixes for a newer Android build.

 * Switches to c++_static STL
 * Builds for all ABIs
 * Chooses android-26 as a recent option
 * Adds a build script for Android
 * Updates libjpeg-turbo build

The build script creates an APK that can be installed with:

  cd android
  ./build
  adb install glmark2.apk

SIMD is disabled in libjpeg-turbo since it simplifies the build
and isn't used in any benchmark update loops. An explicit source
file list is necessary since some sources include other c files.